Move sqlite/spatialite system tables to the QgsSqliteUtils class
